requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
tensorflow>=2.13
wfdb>=4.1
joblib>=1.3
PyYAML>=6.0
flask>=3.0
flask-cors>=4.0
tqdm>=4.66
jupyter>=1.0
notebook>=7.0
ipykernel>=6.0
